chore: cleanup source_notify_button experiment it won
