+++ packages/shared/src/components/sources/SourceActions/SourceActionsBlock.tsx
@@ -1,8 +1,6 @@
 import React, { ReactElement } from 'react';
 import { Button, ButtonSize, ButtonVariant } from '../../buttons/Button';
-import { BlockIcon, PlusIcon } from '../../icons';
-import { useFeature } from '../../GrowthBookProvider';
-import { feature } from '../../../lib/featureManagement';
+import { BlockIcon } from '../../icons';
 
 interface SourceActionBlockProps {
   isBlocked: boolean;
@@ -13,10 +11,8 @@ const SourceActionsBlock = ({
   isBlocked,
   onClick,
 }: SourceActionBlockProps): ReactElement => {
-  const isNotifyExperiment = useFeature(feature.sourceNotifyButton);
-  const unblockLabel = isNotifyExperiment ? 'Unblock' : 'Follow';
-  const defaultIcon = isBlocked ? <PlusIcon /> : <BlockIcon />;
-  const icon = isNotifyExperiment ? <BlockIcon /> : defaultIcon;
+  const unblockLabel = 'Unblock';
+  const icon = <BlockIcon />;
   const label = isBlocked ? unblockLabel : 'Block';
 
   return (
